requirements.txt
Flask==3.0.0
pymongo==4.6.0
pandas==2.1.4
openpyxl==3.1.2
qrcode[pil]==7.4.2
Pillow==10.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-WTF==1.2.1
WTForms==3.1.1
gunicorn
certifi